Pruning step #1
Strips everything that isn't related to Qt5 or at least indirectly to
KDE. For now I've left in both KDE4 and KF5; those could be moved to a
branch if they get in the way and interfere with a clear overview.

Note the 2 previous commits which upgrade mlt and qgit.

Committed from host : Portia.local
Committing to working copy : macstrop-mkae